Break out homepage into components
