+++ stories/home-news-section.stories.tsx
@@ -1,7 +1,6 @@
 import HomeNewsSection from "@/components/home-news-section";
 import { Meta, StoryObj } from "@storybook/react";
 import { faker } from "@faker-js/faker";
-import Container from "@/components/container";
 
 const meta: Meta<typeof HomeNewsSection> = {
   title: "Home/Home News Section",
@@ -56,12 +55,6 @@ type Story = StoryObj<typeof HomeNewsSection>;
 
 export const Default: Story = {
   render({ heroPost, extraPosts }) {
-    return (
-      <div className="p-4 lg:p-8 bg-white dark:bg-stone-900">
-        <Container>
-          <HomeNewsSection heroPost={heroPost} extraPosts={extraPosts} />
-        </Container>
-      </div>
-    );
+    return <HomeNewsSection heroPost={heroPost} extraPosts={extraPosts} />;
   },
 };
